[INLONG-7567][Sort] Extract metrics as common parameters
