+++ inlong-sort/sort-connectors/cdc-base/src/main/java/org/apache/inlong/sort/cdc/base/source/IncrementalSource.java
@@ -38,7 +38,7 @@ import org.apache.flink.metrics.MetricGroup;
 import org.apache.flink.util.FlinkRuntimeException;
 import org.apache.inlong.sort.base.metric.MetricOption;
 import org.apache.inlong.sort.base.metric.MetricOption.RegisteredMetric;
-import org.apache.inlong.sort.cdc.base.config.JdbcSourceConfig;
+import org.apache.inlong.sort.cdc.base.config.BaseSourceConfig;
 import org.apache.inlong.sort.cdc.base.config.SourceConfig;
 import org.apache.inlong.sort.cdc.base.debezium.DebeziumDeserializationSchema;
 import org.apache.inlong.sort.cdc.base.dialect.DataSourceDialect;
@@ -109,7 +109,7 @@ public class IncrementalSource<T, C extends SourceConfig>
             throws Exception {
         // create source config for the given subtask (e.g. unique server id)
         C sourceConfig = configFactory.create(readerContext.getIndexOfSubtask());
-        JdbcSourceConfig jdbcSourceConfig = (JdbcSourceConfig) sourceConfig;
+        BaseSourceConfig baseSourceConfig = (BaseSourceConfig) sourceConfig;
         FutureCompletingBlockingQueue<RecordsWithSplitIds<SourceRecords>> elementsQueue =
                 new FutureCompletingBlockingQueue<>();
 
@@ -121,8 +121,8 @@ public class IncrementalSource<T, C extends SourceConfig>
 
         // create source config for the given subtask (e.g. unique server id)
         MetricOption metricOption = MetricOption.builder()
-                .withInlongLabels(jdbcSourceConfig.getInlongMetric())
-                .withAuditAddress(jdbcSourceConfig.getInlongAudit())
+                .withInlongLabels(baseSourceConfig.getInlongMetric())
+                .withAuditAddress(baseSourceConfig.getInlongAudit())
                 .withRegisterMetric(RegisteredMetric.ALL)
                 .build();
